Fix namespace model extensions.
