+++ TopModel.Core/Model/ClassMappings.cs
@@ -1,5 +1,4 @@
 ﻿using TopModel.Core.FileModel;
-using TopModel.Core.Utils;
 
 namespace TopModel.Core.Model;
