+++ TopModel.Generator.Jpa/ClassGeneration/JavaEnumConstructorGenerator.cs
@@ -1,5 +1,4 @@
 ﻿using TopModel.Core.Model;
-using TopModel.Core.Utils;
 
 namespace TopModel.Generator.Jpa.ClassGeneration;
