+++ TopModel.Generator.Jpa/JpaConfig.cs
@@ -2,7 +2,6 @@
 using TopModel.Core.FileModel;
 using TopModel.Core.Model;
 using TopModel.Core.Model.Implementation;
-using TopModel.Core.Utils;
 using TopModel.Generator.Core;
 using TopModel.Utils;
