+++ TopModel.Generator.Sql/SqlConfig.cs
@@ -1,7 +1,6 @@
 ﻿using TopModel.Core;
 using TopModel.Core.FileModel;
 using TopModel.Core.Model;
-using TopModel.Core.Utils;
 using TopModel.Generator.Core;
 using TopModel.Generator.Sql.Procedural;
 using TopModel.Generator.Sql.Ssdt;
